requirements.txt
discord.py
python-dotenv
mkdocs
